Python3 compatibility and batchnormalization mode
I am using python3 while this is developed based on python2. In python 3, everything becomes float after dividing operation is conducted so int should be applied again if it is calculated before using it as a variable specified as int. In python 2, int becomes int again even if the divide operator is performed between two integers. Hence, int() operation was no needed. Now, it is inconvenient sometimes but this approach changed in python3 is more relevant to our understanding. 
For batchnormalization, I didn't understand clearly how mode=1 and mode=2 are different yet. However, the python result recommends me to use mode=2 so I changed it in discriminator modeling part. It worked after it is changed. Later, I will try to reveal what was a real problem in this problem. Now, I don't do that anything since coding is working anyway.